raspberry_pi/requirements.txt
picamera2
RPi.GPIO
Flask